### 2018-11-12 #36 @whedon
[PRE REVIEW]: Python for Atmosphere and Ocean Scientists
**Submitting author:** @DamienIrving (<a href="http://orcid.org/0000-0003-1258-5002">Damien Irving</a>)
**Repository:** <a href="https://github.com/carpentrieslab/python-aos-lesson" target ="_blank">https://github.com/carpentrieslab/python-aos-lesson</a>
**Version:** v1.0.0
