### 2018-11-12 #36 @labarba
👋  @willingc  — I've assigned this submission to you as handling editor. Since this will be your first editing assignment, let me know if you need to bounce ideas about selecting reviewers. Check out the [Editor's Guide](https://github.com/labarba/jose/blob/master/EditorsGuide.md) I put together: it describes the editor workflow, including needed `@whedon` commands.
