### 2018-11-14 #36 @DamienIrving
In terms of suggested reviewers, the following people have a strong background/interest in educating atmosphere and ocean scientists on best practices for python programming (and data science more generally): @darothen, @RobTheOceanographer, @nicolasfauchereau

@lwasser might also be a potential reviewer, as she has recently been involved in developing geospatial Data Carpentry lessons (that are aimed at a slightly different audience than my lessons)
