### 2018-11-20 #36 @labarba
Thank you, @darothen !
I will now add you as reviewer for this submission, but stay tuned while we look for a second reviewer. In the meantime, have a look at the [Reviewer Guidelines](http://jose.theoj.org/about#reviewer_guidelines) and let me know if you have any questions.
