### 2018-11-20 #36 @labarba
Once we start the review, that will start a new issue thread, and you will have a review checklist to go through, while you enter comments in the thread and possibly open issues on the submission repository. For example, have a look at a review concluded recently: https://github.com/openjournals/jose-reviews/issues/29
