### 2018-11-21 #37 @whedon
[REVIEW]: Python for Atmosphere and Ocean Scientists
**Submitting author:** @DamienIrving (<a href="http://orcid.org/0000-0003-1258-5002">Damien Irving</a>)
**Repository:** <a href="https://github.com/carpentrieslab/python-aos-lesson" target ="_blank">https://github.com/carpentrieslab/python-aos-lesson</a>
**Version:** v1.0.0

- [x] **Version:** Does the release version given match the repository release (v1.0.0)?

- [x] **Version:** Does the release version given match the repository release (v1.0.0)?
