### 2018-11-21 #37 @darothen
@labarba Thanks! I'll take a first pass through things over the holiday weekend and let you know if I have any questions. My goal is to be expeditious here, but could you provide an idea of deadlines you'd like us to hit for wrapping this up?
