### 2018-11-22 #37 @labarba
The ideal situation is a review lasting two or three weeks.
