### 2018-11-25 carpentries-lab/python-aos-lesson#14 @darothen
[JOSE] Release versioning to match JOSE publication
Submission to JOSE is indicated as v1.0.0, but the most recent release tag is v9.3.1. I think tweaks could be made in either place (here or in the JOSE submission); if a new v1.0.0 release is cut here, it would be helpful to add a note to the `README.md` indicating how the sequence of version numbers diverges.

CC: openjournals/jose-reviews#37
