### 2018-11-25 carpentries-lab/python-aos-lesson#15 @darothen
[JOSE] Update AUTHORS file
[AUTHORS](https://github.com/carpentrieslab/python-aos-lesson/blob/gh-pages/AUTHORS) should include @DamienIrving to synchronize with JOSE submission.

CC: openjournals/jose-reviews/issues/37
