### 2018-11-25 carpentrieslab/python-aos-lesson#16 @darothen
[JOSE] Include reference to consolidated, finished product
In the [GitHub](https://github.com/carpentrieslab/python-aos-lesson/edit/gh-pages/_episodes/06-github.md) section of the lesson, the author uses screenshots of a repository hosted at https://github.com/DamienIrving/data-carpentry to walk the student through the process of getting started on GitHub. Unfortunately, that repository doesn't actually exist - or at least, it's not publicly accessible! There's an opportunity here to improve the material in two small ways:

1. Go ahead and create this repository, or make it publicly accessible. Use it as a "living record" of the materials from the lesson, with all the files, commits, changes, and whatnot explicitly included therein.
2. Add a final "Summary" step to the lesson which reviews the "Key Points" listed at the conclusion of each module, and any other useful summary statements. This final step can explicitly linked to the reference GitHub repository which contains the finished product, which the student can use as a reference in case they wish to experiment further.

CC: openjournals/jose-reviews/issues/37
