### 2018-11-25 carpentrieslab/python-aos-lesson#18 @darothen
[JOSE] Module description in JOSE paper
The [paper for submission to JOSE](https://github.com/carpentrieslab/python-aos-lesson/blob/gh-pages/paper.md) is missing an explicit short description of what, exactly the learning materials are and the sequence that they would be introduced to a student. This is partially covered in the "Summary" section, but could really be broken down into its own section, which should include:

- a list of the core skills introduced to the student
- information about the sequence at which the student will encounter the new skills
- examples of tangibles produced during the lesson (e.g. a Figure of one of the outputs from the precipitation plotting script, or a Listing with some code from that script)

CC: carpentrieslab/python-aos-lesson/issues/17
