### 2018-11-25 #37 @darothen
# Review / genereal comments

The materials presented here are a valuable contribution to the atmospheric/oceanic/climate science communities. They succinctly and simply cover a complete coding exercise which most students in the field will encounter routinely, and emphasizes best practices both with regards to scientific Python (e.g. using the "right tools" for the job), coding (using functions, building simple command line programs), software engineering (how to structure a re-usable piece of code, defensive coding), and development (using version control, re-factoring once you have an MWE). Furthermore, the materials are laid out in a logical way that most experienced coders in the field should easily be able to use to teach the material to their peers.

Pending a few small changes to the manuscript and supporting material which will address some of the JOSE requirements and improve their overall quality, I strongly encourage publication in JOSE.

# Review pending checklist / specific comments

## General Checks

- [ ] **Version** - carpentrieslab/python-aos-lesson/issues/14
- [ ] **Authorship** - carpentrieslab/python-aos-lesson/issues/15

## Documentation

- As a general comment here, the compiled module has good documentation for all of these topics, it's just not explicitly set out like in the JOSE paper. I think that's fine; at no point while reviewing the module did I feel lost or not understand why I was doing what I was doing.
- I unfortunately did not have access to a Windows machine to easily review the Windows setup instructions.
- For **Usage**, since this is a set of self-contained educational materials and documentation, I evaluated based on the ease of which it seemed to engage in a "self-guided" manner. I don't think  there are any issues here, but it couldn't hurt to add a sentence or two to the index page of the module which indicates that a user can run through the material at their own pace, outside of a Data Carpentry event.

## Pedagogy / Instructional design

Overall, this is a really well put together set of instructional materials. To directly respond to the author's request for comments, I've assembled the following short list of issues which will hopefully provide some recommendations for small improvements which will boost the overall utility and clarity of the instructional materials.

- [ ] carpentrieslab/python-aos-lesson/issues/16
- [ ] carpentrieslab/python-aos-lesson/issues/17

In the spirit of the JOSE paper, these issues may make good opportunities for collaborators to get involved with the core project and help flesh out and improve it. Triaging these issues as targets for new contributors would be an excellent way to address them.

## JOSE Paper

- [ ] carpentrieslab/python-aos-lesson/issues/18
