### 2018-12-02 carpentrieslab/python-aos-lesson#16 @DamienIrving
Thanks for this suggestion, @darothen. In the past I have deleted the https://github.com/DamienIrving/data-carpentry repo at the conclusion of any workshop I've run, but I can see the value in leaving it up. Next time I teach the workshop I'll make sure that I don't delete it.
