### 2018-12-02 3b95489
Edit to make contents of materials explicit
