### 2018-12-03 carpentrieslab/python-aos-lesson#14 @DamienIrving
Thanks for pointing out this inconsistency, @darothen. I haven't actually tagged any releases of the lessons - those previous releases relate to the development of the [Carpentries lesson template](https://github.com/carpentries/lesson-example) and were inherited when I used GitHub's importer to create my own copy.

I'm thinking that once I've addressed reviewer comments and the JOSE paper is accepted for publication, I'll create a new v1.0.0 tag (adding a note about disregarding earlier version numbers).
