### 2018-12-03 carpentrieslab/python-aos-lesson#14 @darothen
@DamienIrving that sounds like a perfect plan!
