### 2018-12-03 dfac774
Added realm check
