+++ code/plot_precipitation_climatology_final.py
@@ -36,7 +36,7 @@ def apply_mask(darray, sftlf_file, realm):
     """
    
     dset = xr.open_dataset(sftlf_file)
-   
+    assert realm in ['land', 'ocean'], """Valid realms are 'land' or 'ocean'"""
     if realm == 'land':
         masked_darray = darray.where(dset['sftlf'].data < 50)
     else:
