### 2018-12-03 5a3f39d
Moved pdb to defensive programming lesson
