+++ episodes/08-defensive.md
@@ -7,13 +7,12 @@ questions:
 objectives:
 - "Explain what an assertion is."
 - "Add assertions that check the program's state is correct."
-- "Correctly add precondition and postcondition assertions to functions."
+- "Debug Python scripts using the `pdb` library."
 - "Identify sources of more advanced lessons on code testing."
 keypoints:
 - "Program defensively, i.e., assume that errors are going to arise, and write code to detect them when they do."
 - "Put assertions in programs to check their state as they run, and to help readers understand how those programs are supposed to work."
-- "Use preconditions to check that the inputs to a function are safe to use."
-- "Use postconditions to check that the output from a function is safe to use."
+- "The `pdb` library can be used to debug a Python script by stepping through line-by-line."
 - "Software Carpentry has more advanced lessons on code testing."
 ---
 
@@ -44,8 +43,9 @@ Python does nothing,
 but if it's false,
 Python halts the program immediately
 and prints the error message if one is provided.
-For example,
-this piece of code halts as soon as the loop encounters a value that isn't positive:
+
+To demonstrate an assertion in action,
+consider this piece of code that halts as soon as the loop encounters a value that isn't positive:
 
 ~~~
 numbers = [1.5, 2.3, 0.7, -0.001, 4.4]
@@ -74,152 +74,157 @@ AssertionError: Data should only contain positive values
 Programs like the Firefox browser are full of assertions:
 10-20% of the code they contain
 are there to check that the other 80-90% are working correctly.
-Broadly speaking,
-assertions fall into three categories:
 
-*   A *precondition* is something that must be true at the start of a function in order for it to work correctly.
-
-*   A *postcondition* is something that the function guarantees is true when it finishes.
-
-*   An *invariant* is something that is always true at a particular point inside a piece of code.
-
-For example,
-suppose we are representing rectangles using a tuple of four coordinates `(x0, y0, x1, y1)`,
-representing the lower left and upper right corners of the rectangle.
-In order to do some calculations,
-we need to normalize the rectangle so that the lower left corner is at the origin
-and the longest side is 1.0 units long.
-
-![Normalizing a rectangle](../fig/08-defensive-normalize.png)
-
-This function does that,
-but checks that its input is correctly formatted and that its result makes sense:
+To see how assertions might be useful
+in the context of the `plot_precipitation_climatology.py` script,
+let's try the following at the command line:
 
 ~~~
-def normalize_rectangle(rect):
-    '''Normalizes a rectangle so that it is at the origin and 1.0 units long on its longest axis.
-    Input should be of the format (x0, y0, x1, y1). 
-    (x0, y0) and (x1, y1) define the lower left and upper right corners of the rectangle, respectively.'''
-    assert len(rect) == 4, 'Rectangles must contain 4 coordinates'
-    x0, y0, x1, y1 = rect
-    assert x0 < x1, 'Invalid X coordinates'
-    assert y0 < y1, 'Invalid Y coordinates'
[…]
