### 2018-12-03 a21a6ae
Moved pdb to defensive programming lesson
