+++ _episodes/09-provenance.md
@@ -146,7 +146,7 @@ print(new_record)
 >     """
 >   
 >     dset = xr.open_dataset(sftlf_file)
->   
+>     assert realm in ['land', 'ocean'], """Valid realms are 'land' or 'ocean'"""
 >     if realm == 'land':
 >         masked_darray = darray.where(dset['sftlf'].data < 50)
 >     else:
