### 2018-12-03 carpentrieslab/python-aos-lesson#17 @DamienIrving
Thanks, @darothen - this is an excellent suggestion.

I've gone ahead and moved the `pdb` content to the lesson on defensive programming (see https://github.com/carpentrieslab/python-aos-lesson/commit/a21a6ae2e0993999b9cfc3e815e147d1f0121766).
