### 2018-12-05 carpentrieslab/python-aos-lesson#20 @RobTheOceanographer
[JOSE] Add netcdf4 to the install in lesson 01-conda
The xr.open_dataset call in lesson `02-visualisation` threw an TypeError for me. The error was:
```...
TypeError: Error: .../data-carpentry/data/pr_Amon_ACCESS1-3_historical_r1i1p1_200101-200512.nc is not a valid NetCDF 3 file
            If this is a NetCDF4 file, you may need to install the
            netcdf4 library, e.g.,

            $ pip install netcdf4
```
This is likely due to the an assumption that the conda env built by the users will have the python netcdf4 in it... but that was not the case for my set up so I have submitted an issue to add netcdf4 to the install process in lesson `01-conda`.

I recommend adding netCDF4 to to the install explicitly.

CC: openjournals/jose-reviews#37
