### 2018-12-05 carpentrieslab/python-aos-lesson#21 @RobTheOceanographer
[JOSE]  general inconsistencies in how you refer to relative paths 
There are general inconsistencies in how you refer to relative paths - sometimes `data/...` other times `.../data/...`. While most users can figure this out it would be better to go through the lessons and tidy this up

CC: openjournals/jose-reviews#37
