### 2018-12-05 carpentrieslab/python-aos-lesson#22 @RobTheOceanographer
[JOSE] `code` directory that does not exist 
In lesson `04-cmdline` you refer to a `code` directory that does not exist if one has followed the install/setup instructions provided. Best to either ask users to put the .py scripts into a code dir or to refactor the paths to not refer to `code/...`.
 e.g. in lesson 04-cmdline the line `$ cat code/script_template.py` would become `$ cat script_template.py`.

CC: openjournals/jose-reviews#37
