### 2018-12-05 carpentrieslab/python-aos-lesson#23 @RobTheOceanographer
[JOSE] xarray module is imported as `xr` but tied to call as `xarray`
In lesson `07-vectorisation` the first call to xarray.open_dataset has a syntax error as the xarray module is imported as `xr`:
Change `xarray.open_dataset` to `xr.open_dataset` ?

CC: openjournals/jose-reviews#37
