### 2018-12-05 carpentrieslab/python-aos-lesson#23 @RobTheOceanographer
there is actually a second instance of this further down in this lesson that need addressing too.
