### 2018-12-06 #37 @RobTheOceanographer
# Review / general comments
Daniel's review picked up much of what I saw - great work @darothen! - nevertheless, I have some general comments and suggestions to contribute.

The atmospheric and oceanographic science communities are crying out for this type of training material. I am so glad that there are brilliant people like @DamienIrving who are going the extra distance to help bring them up to standard and to educate their peers and colleagues. The materials Damien present here are succinct and well thought out - it is a complete end-to-end experience for the students.

Pending the changes suggested by Daniel and a couple of suggestions I have below, I strongly recommend this manuscript for publication in JOSE.

# Review pending checklist
The version doesn't match yet, but @darothen submitted an issue for it already: carpentrieslab/python-aos-lesson/issues/14

# Lessons
## 02-visualisation
The xr.open_dataset call threw an TypeError for me. This is likely due to  an assumption that the `conda env` built by the users has the python netcdf4 package in it... but that was not the case for my set up, so I have submitted an issue to add netcdf4 to the install process in lesson `01-conda`.

- [ ] carpentrieslab/python-aos-lesson/issues/20

There are general inconsistencies in how you refer to relative paths - sometimes `data/...` other times `.../data/...`. While most users can out what you mean it would be better to go through the lessons and tidy this up:

- [ ] carpentrieslab/python-aos-lesson/issues/21

In lesson `04-cmdline` you refer to a `code` directory that does not exist if you follow the install/setup instructions provided. best to either ask users to put the .py scripts into a code dir or to refactor the paths to not refer to `code/...`. e.g. in lesson 04-cmdline the line `$ cat code/script_template.py` should become `$ cat script_template.py`. :

- [ ] carpentrieslab/python-aos-lesson/issues/22

In lesson `07-vectorisation` the first call to xarray.open_dataset has a syntax error as the xarray module is imported as `xr`:

- [ ] carpentrieslab/python-aos-lesson/issues/23

I LOVE LESSON 08! This lesson must become mandatory for all science postgraduate students.

# Pedagogy
The structure and pedagogy here are largely modelled on the software carpentry and data carpentry style. That is good because it follows a tried and true formula that has been proven to engage students and produce outcomes over many years.

The JOSE reviews guidelines say something about "The authors should briefly explain their 'Instructional design of the module' in the JOSE paper." I'm not sure that this is needed in this case as it is deeply linked to the `carpentrieslab` stuff.
