### 2018-12-06 9cc6f69
Added netCDF4 to installation instructions
