### 2018-12-06 42a1143
Added netCDF4 to installation instructions
