+++ episodes/01-conda.md
@@ -120,6 +120,7 @@ because mixing packages from multiple channels can cause headaches like binary i
 For these particular lessons we will use xarray,
 but all the same tasks could be performed with iris.
 We'll also install
+[netCDF4](http://unidata.github.io/netcdf4-python/) (xarray requires this to read netCDF files),
 [cartopy](http://scitools.org.uk/cartopy/) (to help with geographic plot projections),
 [cmocean](http://matplotlib.org/cmocean/) (for nice color palettes)
 and [jupyter](https://jupyter.org/) (so we can use the jupyter notebook).  
@@ -127,12 +128,12 @@ and [jupyter](https://jupyter.org/) (so we can use the jupyter notebook).
 We could install these libraries from Anaconda Navigator (not shown)
 or using the Bash Shell or Anaconda Prompt (Windows):
 ~~~
-$ conda install jupyter xarray cartopy cmocean
+$ conda install jupyter xarray netCDF4 cartopy cmocean
 ~~~
 {: .language-bash}
 
 If we then list all the libraries that we've got installed,
-we can see that jupyter, xarray, cartopy and cmocean (and their dependencies)
+we can see that jupyter, xarray, netCDF4, cartopy and cmocean (and their dependencies)
 are now there:
 ~~~
 $ conda list
@@ -156,7 +157,7 @@ $ conda list
 > of the Anaconda Navigator or via the following Bash Shell / Anaconda Prompt commands:
 >
 > ~~~
-> $ conda create -n pyaos-lesson jupyter xarray cartopy cmocean
+> $ conda create -n pyaos-lesson jupyter xarray netCDF4 cartopy cmocean
 > $ conda activate pyaos-lesson
 > ~~~
 > {: .language-bash}
