### 2018-12-06 carpentrieslab/python-aos-lesson#20 @DamienIrving
Thanks for spotting this omission, @RobTheOceanographer.

It's frustrating that the conda recipe for xarray doesn't install netCDF4 by default, but indeed at the [xarray website](http://xarray.pydata.org/en/stable/installing.html) it asks people to install netCDF4, bottleneck and dask. These lessons don't cover parallel processing (and hence dask and bottleneck aren't required), but I've gone ahead and added netCDF4 to the install instructions.

See https://github.com/carpentrieslab/python-aos-lesson/commit/9cc6f698e9869f04f6f9911155c58a3a2daf4431
