### 2018-12-06 a22df1d
Fixed path to data directory
