+++ _episodes/03-functions.md
@@ -30,7 +30,7 @@ import cartopy.crs as ccrs
 import numpy as np
 import cmocean
 
-access_pr_file = '../data/pr_Amon_ACCESS1-3_historical_r1i1p1_200101-200512.nc'
+access_pr_file = 'data/pr_Amon_ACCESS1-3_historical_r1i1p1_200101-200512.nc'
 
 dset = xr.open_dataset(access_pr_file)
 
