### 2018-12-06 3e1ca5a
Used xr abbreviation for xarray
