+++ episodes/07-vectorisation.md
@@ -21,7 +21,7 @@ import xarray as xr
 
 access_sftlf_file = 'data/sftlf_fx_ACCESS1-3_historical_r0i0p0.nc'
 
-dset = xarray.open_dataset(access_sftlf_file)
+dset = xr.open_dataset(access_sftlf_file)
 sftlf = dset['sftlf']
 print(sftlf)
 ~~~
@@ -73,7 +73,7 @@ it is a land point (and thus needs to be set to `np.nan`).
 (For this example, we are going to define land as any grid point that is more than 50% land.)
 
 ~~~
-dset = xarray.open_dataset('data/pr_Amon_ACCESS1-3_historical_r1i1p1_200101-200512.nc')
+dset = xr.open_dataset('data/pr_Amon_ACCESS1-3_historical_r1i1p1_200101-200512.nc')
 clim = dset['pr'].mean('time', keep_attrs=True)
 
 nlats, nlons = clim.data.shape
