### 2018-12-06 carpentrieslab/python-aos-lesson#21 @DamienIrving
Thanks for spotting this error, @RobTheOceanographer.

I've gone ahead and corrected all references to the `data` directory (see https://github.com/carpentrieslab/python-aos-lesson/commit/7265c43294f1cbd8da275ca03ad8296721ed801a).
