### 2018-12-06 carpentrieslab/python-aos-lesson#18 @DamienIrving
Thanks, @darothen. This is an excellent suggest. I've gone ahead and editing the text of the paper accordingly (see #19).
