### 2018-12-06 #37 @DamienIrving
Thanks, @darothen and @RobTheOceanographer. It's been really useful to have you both take a detailed look over the lessons. I've gone through and addressed each of the issues you identified.

The following issues were simple bug fixes (each issue links to the relevant commit that addresses the problem):
* https://github.com/carpentrieslab/python-aos-lesson/issues/15
* https://github.com/carpentrieslab/python-aos-lesson/issues/20
* https://github.com/carpentrieslab/python-aos-lesson/issues/21
* https://github.com/carpentrieslab/python-aos-lesson/issues/22
* https://github.com/carpentrieslab/python-aos-lesson/issues/23

One issue resulted in a substantial rewrite of the defensive programming lesson, which I think is now much improved:
* https://github.com/carpentrieslab/python-aos-lesson/issues/17

One issue is a simple thing that I'll do next time I teach the materials (I'll leave the repo created during the GitHub lesson live on the web):
* https://github.com/carpentrieslab/python-aos-lesson/issues/16

I'll fix the issue with the versioning as soon as the lessons are accepted for publication with JOSE (i.e. that will be release 1.0.0):
* https://github.com/carpentrieslab/python-aos-lesson/issues/14

One issue resulted in a rewrite of the second paragraph of the JOSE paper, which is currently in the form of a pull request that I can merge if @darothen is satisfied with the changes:
* https://github.com/carpentrieslab/python-aos-lesson/issues/18
