### 2018-12-08 carpentrieslab/python-aos-lesson#18 @darothen
Hey @DamienIrving, I think the edited text is great! I'll add a note in /jose-reviews/issues/37 confirming this.
