### 2019-01-01 #37 @darothen
@labarba - oh, sorry about that! I strongly recommend publication in the manuscript/materials' current form.
