### 2019-01-03 #37 @DamienIrving
The proofs look good to me. There was just one typo that I fixed up: https://github.com/carpentrieslab/python-aos-lesson/commit/8828c78fff2efeed2b0416a15a83c2a4c2649524
