### 2019-01-18 #37 @DamienIrving
@labarba Now that I've checked the proofs (I fixed a minor typo as a result), is there anything else I need to do?
