### 2019-01-19 #37 @labarba
Thanks for the ping, @DamienIrving !

Please make an archive in Zenodo and post the DOI here. Make sure to edit the metadata (title and author list) to match your JOSS paper. Thanks!
