### 2019-01-19 #37 @labarba
Minor edits:

¶ 1 — comma after e.g.
¶ 2 — "it is also possible work through" > missing preposition "to"
comma after i.e.
(same at start of page 2, and ¶ 2 of page 2, and last paragraph)
period after etc
