### 2019-01-20 b83e91a
Minor fixes
