+++ README.md
@@ -1,10 +1,12 @@
+[![DOI](https://zenodo.org/badge/117483226.svg)](https://zenodo.org/badge/latestdoi/117483226)
+
 # python-aos-lesson
 
 The repository contains the lesson materials for a single day workshop
 on using python (and git) in the atmosphere and ocean sciences:  
 [https://carpentrieslab.github.io/python-aos-lesson/](https://carpentrieslab.github.io/python-aos-lesson/)
 
-The lesson has been written by [Damien Irving](https://twitter.com/DrClimate) based on experiences
+The lesson materials were written by [Damien Irving](https://twitter.com/DrClimate) based on experiences
 from teaching generic Software Carpentry workshops at the annual conference of the
 Australian Meteorological and Oceanographic Society (AMOS) from 2014-2017.
 
@@ -12,3 +14,4 @@ The lesson materials have been used in the following workshops and university co
 * [Data Carpentry workshop at the 2018 AMOS Conference](https://damienirving.github.io/2018-02-04-amos-icshmo/), February 2018
 * [Data Carpentry workshop at Woods Hole Oceanographic Institution](https://damienirving.github.io/2018-08-15-whoi/), August 2018
 * Changing Global Environments, 3rd year undergraduate subject, University of Wollongong, Semester 2, 2018
+* [Data Carpentry workshop at the 2019 AMOS Conference](https://damienirving.github.io/2019-06-10-amos/)
