### 2019-01-21 17ae9ee
Added JOSE badge
