### 2019-01-21 #37 @labarba
Yay! Congratulations, @DamienIrving, your paper is published in JOSE and the DOI is active.

A big thank you to your editor, @willingc,  and your reviewers: @darothen, @RobTheOceanographer 🙏
