### 2019-01-21 #37 @DamienIrving
Thanks @willingc, @darothen, @RobTheOceanographer for the review and also @labarba for your efforts in establishing JOSE.
