### 2019-03-21 fc55f51
Added citation
