+++ README.md
@@ -15,4 +15,5 @@ The lesson materials have been used in the following workshops and university co
 * [Data Carpentry workshop at the 2018 AMOS Conference](https://damienirving.github.io/2018-02-04-amos-icshmo/), February 2018
 * [Data Carpentry workshop at Woods Hole Oceanographic Institution](https://damienirving.github.io/2018-08-15-whoi/), August 2018
 * Changing Global Environments, 3rd year undergraduate subject, University of Wollongong, Semester 2, 2018
+* [Data Carpentry workshop at the Australian Bureau of Meteorology](https://damienirving.github.io/2019-05-23-bom/), May 2019
 * [Data Carpentry workshop at the 2019 AMOS Conference](https://damienirving.github.io/2019-06-10-amos/), June 2019
